+++ SuperMarioBros3/GreenKoopas.cpp
@@ -127,7 +127,7 @@ void GreenKoopas::Render()
 		break;
 	}
 
-	CAnimations::GetInstance()->Get(ani)->GetTransform()->Scale = Vec2(-nx, transformation.y);
+	CAnimations::GetInstance()->Get(ani)->GetTransform()->Scale = Vec2((float)-nx, transformation.y);
 	CAnimations::GetInstance()->Get(ani)->Render(position.x, position.y);
 	RenderBoundingBox();
 }
